kalakriti-cms-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-jose
pydantic
email-validator
python-multipart
boto3